pyproject.toml
[project]
name = "llamatest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.16",
    "nvidia-ml-py>=13.590.44",
    "psutil>=7.2.1",
]
